requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-rest-auth==7.0.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mysqlclient==2.2.6
paypalrestsdk==1.13.3
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
urllib3==2.2.3
